Scope WebdriverIO's global types to the iOS tests
The root tsconfig listed @wdio/globals/types, @wdio/mocha-framework, and
@wdio/browserstack-service under compilerOptions.types, so their ambient
declarations applied to every file in src: @wdio/browserstack-service's global
State, GRRUrls, and TestContextOptions interfaces, WebdriverIO's
browser/$/$$/driver/expect, and mocha's describe/it. A file that forgot to
import the app's own State compiled silently against BrowserStack's.

Give the iOS WebdriverIO tests their own tsconfig (src/e2e/iOS/tsconfig.json),
the only program that declares those globals, and exclude src/e2e/iOS from the
root program. Move the WebdriverIO half of the console proxy out of
src/util/consoleProxy.ts, the one app module that used the browser global,
into wdio.base.conf.ts, its only consumer. Route src/e2e/iOS to the new project
in eslint.config.js and run both programs from lint:tsc.
