+++ package.json
@@ -47,7 +47,7 @@
     "lint:no-npm": "node ./scripts/no-files.js package-lock.json",
     "lint:prettier": "FORCE_COLOR=1 prettier --check .",
     "lint:src": "FORCE_COLOR=1 NODE_OPTIONS=--max-old-space-size=4096 eslint . --cache --rule 'no-console: [2, { allow: [error, info, warn] }]'",
-    "lint:tsc": "tsc",
+    "lint:tsc": "tsc && tsc -p src/e2e/iOS/tsconfig.json",
     "lint": "FORCE_COLOR=1 npm-run-all --serial \"lint:*\"",
     "postinstall": "yarn build:packages && git config core.hooksPath .hooks && yarn build:styles",
     "servebuild": "vite preview --port 3000",
